Update AddNode to always add to root if no parent specified
Previously, AddNode would add a new node under the selected node if no explicit parent was provided. Now, it always adds to the root when no parent is specified. This removes one implicit scenario, the user has to specify if they want to add to currently selected node by passing it in as parent.